pyproject.toml
[project]
name = "hevy"
version = "0.1.0"
description = "Hevy Workout App MCP."
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "gumstack-mcp==1.23.3+gumloop.0.1.33",
    "httpx>=0.27.0",
    "pydantic>=2.13.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
hevy = "hevy.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hevy"]

[tool.uv]
package = true

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[[tool.uv.index]]
name = "gumstack"
url = "https://us-west1-python.pkg.dev/gumstack-public/gumstack-mcp/simple/"
explicit = true

[tool.uv.sources]
gumstack-mcp = { index = "gumstack" }
